+++ src/web/010_schemes.js
@@ -193,10 +193,10 @@ local.schemes.register('httpl', function(request, response) {
 		if (peerd) {
 			// See if this is a default stream miss
 			if (peerd.stream === '0') {
-				if (request.urld.authority.slice(-2) == ':0') {
+				if (request.urld.authority.slice(-2) == '!0') {
 					server = local.getServer(request.urld.authority.slice(0,-2));
 				} else {
-					server = local.getServer(request.urld.authority + ':0');
+					server = local.getServer(request.urld.authority + '!0');
 				}
 			}
 			if (!server) {